Retrieval Augmented Generation (RAG)/RAG_Demo/requirements.txt
langchain
langchain-community
langchain-core
langchain-experimental
langchain-chroma
langchain-openai
langchain-ollama
langchain-groq
groq
ollama
openai
sentence-transformers
python-dotenv
streamlit
langchainhub
unstructured
ipykernel